+++ source/classes.cpp
@@ -15,7 +15,7 @@ struct gun{
     int damage;
     int ammo;
     int ammo_max;
-    float fire_rate;
+    int fire_rate;
     void shoot();
 };
